requirements.txt
# Core runtime
pyyaml>=6.0
typer[all]>=0.9

# Optional REST micro-service
fastapi>=0.110
uvicorn[standard]>=0.27

# Type annotations (shipped)
typing-extensions>=4.10